Added collider support
Which is half working, half broken.
You can grab individual items without issues, but due to
the inverse scale at the top of the hierarchy, the grab
distance is broken. I can't set it up to only allow grabbing
at point blank.

Also fixed an issue in the main script.
Added a few signals when the loaded model is removed
and when the model is actually loaded.

Signed-off-by: Voyage <voyage@miouyouyou.fr>