Fixed inverse scale and scheduling loads
The inverse Y scale is no more, thanks to @lyuma2d

I still need to put a 180 rotation on the Y axis of
the root node so that everything aligns in the same
"direction" (Like things going to the right in Blender
goes to the right in Unity).

However, the inverse scale is gone.

And also, the parsing now have a state, a scheduler
and can leave the hand back to the main thread, and
then resume later.

This makes loading HEAVY GLB files a bit easier and
should avoid hogging on the CPU too much.
Still, for every task, the first operation is still
done ANYWAY, so if that operation hogs the CPU, nothing
can be done at the moment.

Signed-off-by: Voyage <voyage@miouyouyou.fr>