PlayCanvas monstrosities support
Well, that forced me to actually access buffers using
accessors, instead of trying to shortcut directly to
the buffer views.

But for fuck sake... Why would you provide a number
of points that is not a multiple of 3 is you're going
to provide triangles... Like, WHY !? It's a TRIANGLE !
It's supposed to have 3 POINTS PLAYCANVAS !

Still, I'm starting to support PlayCanvas GLB correctly.

Also Transparent Cutout is now supported... in the editor
at least !

Signed-off-by: Voyage <voyage@miouyouyou.fr>