BC7 textures and integer indices support
The converter WORKS !
Took me a few days to wrap my head around the basics of
Rust but it WORKS !

I now have a GLB converter that can convert the textures to
BC7.

So I tested it on some random Sketchfab models and, while the
textures were converted correctly, I couldn't display the model
because it used Integer indices while I was always (wrongly)
expecting integer indices.

The problems are solved ! I'll upload the converter too, but
it's not completely ready, though.

Signed-off-by: Voyage <voyage@dontcontactmebymail.com>