Matrices support !
So when trying to load Sketchfab models pre-converted
using a texture converter I wrote in Rust, it turned
out that the model appeared in pieces...

The reason being that, in GLTF, you can define the
transform of a node using either a "position", "rotation"
and "scale".
Or you can just feed AN ENTIRE MATRIX for "REASONS".

This new update handles the provided 4x4 Matrix and use
methods actually provided by Unity to convert TRS matrices
to "Position", "Rotation" and "Scale"

Signed-off-by: Voyage <voyage@dontcontactmebymail.com>