Support for Occlusion Maps
At least, the Map is loaded into the right texture slot.
I still need to find out how to fire "multi_compile" so that
every single variant is compiled correctly and can be loaded
in-game without issues.

Also, a new test suite has been put in place in
https://github.com/vr-voyage/vrchat-glb-loader-tests

This allows me to test the existing code and check for
potential regressions.

Signed-off-by: Voyage <voyage@miouyouyou.fr>