Added support for the "Asset" property
Which is actually REQUIRED in GLTF 2.0.
This property is NEEDED and the parsing must fail
if such property is not present.

Sketchfab uses this property to store information about
the asset author as well, so I might update the UI to
display this information.
Meanwhile, the information is currently stored in a
GLTFAsset object.

Signed-off-by: Voyage <voyage@miouyouyou.fr>