Propert "texture" support
With sampler and images properties parsed correctly this time !
Before hand, I was using a stupid shortcut, that consisted in
taking data from image "4" when requesting texture "4".

However, texture "1", "2", "3", "4" could all be referencing
image "0" actually.

So yeah, now all the textures are based on the informations
parsed from "textures" !

The "samplers" are parsed too and, when creating a Texture,
the properties are applied "as best as possible", since OpenGL
tries to differentiate Minifying Filter and Magnifying Filter
while Unity (and most games engines actually) don't.

So, yeah, the rules here are :
* If you use NEAREST or NEAREST_MIPMAP_NEAREST, you get
  Point Filtering
* If you use LINEAR_MIPMAP_LINEAR, you get Trilinear
* For anything else you get Bilinear

Clamping/Mirror/Repeat are also applied correctly per-axis.

Signed-off-by: Voyage <voyage@miouyouyou.fr>