Separated the 'UI'
Yes, that ugly board is a UI and if I put a script to manage
it, the project suddenly depends on TMPro. Ugh...

So, I separated it the script managing the Assets Info into
a specific folder. I might do the same with anything UI related
after some point, in order to distinguish the core from the
different 'UI' elements.

Speaking about UI, I think it's time this thing ressembles
something a bit more interesting, before a 1.0 release.

Signed-off-by: Voyage <voyage@miouyouyou.fr>