Added MSFT_texture_DDS support
Added support for the Microsoft DDS texture extension.
This leads to the addition of a DDS parser script.

Parsing the DDS header is quite of a pain. You don't have
any information about the amount of data to expect, and
checking for the format require to play with various FourCC
and extensions analysis tricks.

The current implementation is quite clunky but works fine !

Signed-off-by: Voyage <voyage@miouyouyou.fr>