Starting to add support for ShaderMotion
Since I'm out of options when it comes to simple
Unity friendly CPU Skinning, I'm going to the GPU Skinning route.

My best option at the moment is ShaderMotion so I'm trying
to make it work with it.

I started to port the "CreatePlayer" and Mesh Combining algorithm.
It has been tested against the actual official implementation and
provide the same results, with uv1 being APPROXIMATELY the same
and no UV0 for the Blendshapes textures at the moment.

There's still a lot to implement, so #3 is not going to be
resolved soon...

Signed-off-by: Voyage <voyage@miouyouyou.fr>