Starting to add ShaderMotion support
Since we cannot get support for Bones, we'll do
animation another way.

There are multiple ones actually.

The first one I'd like to support is ShaderMotion,
so here it is.
It's still not *ready*, since I need to release the
encoder and, most importantly, make :
* The Mesh Player "MToon" compatible, in order to get
  nice shading (rather easy I think)
* Make the encoder standalone (very tricky)

I think I might do a release and make the encoder
"Unity only" at the beginning, if I can release a MToon-like
ShaderMotion MeshPlayer shader.

Signed-off-by: Voyage <voyage@miouyouyou.fr>