Finally starting to release !
Took 6 months to wrap my head around all the things
needed to release this correctly.

Between :
* the generation of the 3 packages
  (loader, shaders, utils);
* the hell that UPM is at the end of 2025;
* various steps, hoops and loops for putting a up a VCC
  repository that works

I'm glad to see that it's ALMOST over.
This will generate a painful update, though, since most
of the shaders will be moved into another project, that
this project will start to depend on.
This will require people to uninstall the previous
installed GLB Loader and install a new one "hoping things
don't break midway....".

That split-up might actually generate a few issues on the
long-term, but I didn't find other sane solution, given
the lack of real and useful choices at the end of 2025.

Signed-off-by: Voyage <voyage@miouyouyou.fr>