Slight rework of the "UI"
That fugly UI needs to die at some point.
And so the Asset Info are now hidden by default.
The only remaining part is the URL input part, with
the "GLB URL" text being replaced by
"3D Model URL (.glb/.vrm)" which should be easier
to understand.

Still, this REALLY needs to be replaced.

Also removed the Editor folder.

There's a Utilities package now and, the "Editor"
part has no use at the moment.
Creating "Editors" for the GLBLoader asset itself
it not very useful IMHO.
And keeping an empty folder generates useless warnings
so... let's remove it until there's a need for it.

Signed-off-by: Voyage <voyage@miouyouyou.fr>